design the quiz data structure and quiz model that operates on that data
The quiz is multiple choice and mainly consists of an object that wrappers
an array of questions, choices, and correct answer.

The quiz also supports the notion of themes which could entail graphics and
sound design.  Most of this is stubbed out for now.

The choices are also objects (instead of just text), in the event we would
like to support images or giphys in the future.