teach the quiz model about praise, approbation, and timeouts
After each question, we're supposed to offer up some response.
The model now has an array of options to serve up depending if the
question were correctly or incorrectly answered.  This is organized
under the theme of the quiz and could be more or less affirmative or
snarky depending upon the personality you want to convey.

Also, we codify the amount of time in seconds the user has to answer
a given question.  A pause between questions is also provided for use
by the controller.

So pacing of the quiz itself is nominally configurable through the json data
provided when the quiz is instantiated.  I could see the quiz engine itself
accepting mutations of these parameters from the controller as a user-preference.