add ability to click on a choice and have it compared to the answer
With this, /most/ of the controller is basically functional.
Still need to handle some edge cases, like when the quiz completes
because it timed out on the last question.  (Currently, Most of the fu for
recognizing the end of a game is in a click-driven callback.)

Also, still need to report results and ask if they want to replay.