normalize how we accomodate quizzes of various schema
Here, I totally break the app. :-)

But it's in the service of something good, namely the ability to support quiz data from
the OpenTDB endpoint.  Yes, there is such a thing.  You tell it what category of quiz
items you want, and it magically sends you a tasty chunk of json, provided your app knows
how to consume that json.  For now, I'm only supporting statically downloaded instances
of such data, but in a follow-on enhancement, I could see allowing realtime access to the
trivia database.

With this commit, I'm teaching the app how to process OpenTDB input in a scalable way
that leaves open the prospect of supporting other quiz schemas (like Quizlet) some day.

Most of the action of significance here is in assets/js/quiz-factory.js.

Expect some follow-on commits which clean up what I've broken.