add quiz item iteration under timer and click control
This adds the ability to iterate over an array of quiz questions
under timer control.

If the user is too slow to respond, the correct answer is shown
before advancing to the next question.

If the user selects an answer in time, their choice is checked
and congratulations or correction alert is displayed before
advancing to the next quiz item.

I still need to implement the visual 1s count down clock.